backend/requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
werkzeug==2.0.3